requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.15
dnspython==2.6.1
docx==0.2.4
elevenlabs==1.50.3
executing==2.1.0
fastavro==1.10.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.3.1
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
gradio_client==1.5.2
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
ipykernel==6.29.4
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
noise==1.2.2
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.2.0
pooch==1.8.2
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyFLAC==3.0.0
pygame==2.6.1
PyGithub==2.5.0
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyJWT==2.10.1
pymongo==4.9.1
PyNaCl==1.5.0
PyOpenGL==3.1.7
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==25.1.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.27.1
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.12.0
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
trio==0.27.0
trio-websocket==0.11.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.17.0
wsproto==1.2.0
XlsxWriter==3.2.0
